Introduced SocketWriterInitializer.
Idea is not to rely on subclassing otherwise one would need to have
MemberNonBlockingSocketWriter
MemberSpinningSocketWriter
ClientNonBlockongSocketWriter
ClientSpinningSocketWriter

This is just for the writers; the readers have the same issue.

Also it would impose a subclassing approach on the IOTHreadModel leading
to another 4 subclasses (2 models * 2 types)
