+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/SocketWriter.java
@@ -18,6 +18,8 @@ package com.hazelcast.nio.tcp;
 
 import com.hazelcast.nio.OutboundFrame;
 
+import java.nio.ByteBuffer;
+
 /**
  * Each {@link TcpIpConnection} has a {@link SocketWriter} and it writes {@link OutboundFrame} instances to the socket. Copying
  * the Frame instances to the byte-buffer is done using the {@link WriteHandler}.
@@ -85,4 +87,8 @@ public interface SocketWriter {
      * through the {@link TcpIpConnection#close(String, Throwable)} method.
      */
     void close();
+
+    void initOutputBuffer(ByteBuffer outputBuffer);
+
+    void initWriteHandler(WriteHandler writeHandler);
 }
