+++ hazelcast/src/main/java/com/hazelcast/nio/tcp/spinning/SpinningIOThreadingModel.java
@@ -22,6 +22,7 @@ import com.hazelcast.logging.LoggingService;
 import com.hazelcast.nio.tcp.IOThreadingModel;
 import com.hazelcast.nio.tcp.SocketReader;
 import com.hazelcast.nio.tcp.SocketWriter;
+import com.hazelcast.nio.tcp.SocketWriterInitializerImpl;
 import com.hazelcast.nio.tcp.TcpIpConnection;
 
 /**
@@ -45,12 +46,14 @@ public class SpinningIOThreadingModel implements IOThreadingModel<TcpIpConnectio
     private final LoggingService loggingService;
     private final SpinningInputThread inputThread;
     private final SpinningOutputThread outThread;
+    private final SocketWriterInitializerImpl socketWriterInitializer;
 
     public SpinningIOThreadingModel(LoggingService loggingService, HazelcastThreadGroup hazelcastThreadGroup) {
         this.logger = loggingService.getLogger(SpinningIOThreadingModel.class);
         this.loggingService = loggingService;
         this.inputThread = new SpinningInputThread(hazelcastThreadGroup);
         this.outThread = new SpinningOutputThread(hazelcastThreadGroup);
+        this.socketWriterInitializer = new SocketWriterInitializerImpl(logger);
     }
 
     @Override
@@ -61,7 +64,7 @@ public class SpinningIOThreadingModel implements IOThreadingModel<TcpIpConnectio
     @Override
     public SocketWriter newSocketWriter(TcpIpConnection connection) {
         ILogger logger = loggingService.getLogger(SpinningSocketWriter.class);
-        return new SpinningSocketWriter(connection, logger);
+        return new SpinningSocketWriter(connection, logger, socketWriterInitializer);
     }
 
     @Override
